refactor: change breakPoints1 file name to breakpoints to detect file name changes
affects: @medly-components/theme, @medly-components/utils
